refactor(client): :recycle: Rename ThemeDropdown to ThemeMenu
