+++ client/src/components/Pages/homepage/HomeSearchHeader/HomeSearchHeader.desktop.tsx
@@ -1,5 +1,5 @@
 import LocationMenu from "components/Pages/recherche/LocationMenu";
-import ThemeDropdown from "components/Pages/recherche/ThemeDropdown";
+import ThemeDropdown from "components/Pages/recherche/ThemeMenu";
 import EVAIcon from "components/UI/EVAIcon/EVAIcon";
 import { cls } from "lib/classname";
 import { useTranslation } from "next-i18next";
