--- client/src/components/Pages/recherche/ThemeDropdown/index.ts
@@ -1,3 +1,0 @@
-import ThemeDropdown from "./ThemeDropdown";
-
-export default ThemeDropdown;
