+++ client/src/components/Pages/recherche/ThemeMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ThemeMenu";
